Add method to get ChannelMessageCount as flow. Version "1.7.59022-SNAPSHOT"

+++ examples/SceytDemoApp/src/main/java/com/sceyt/chat/demo/presentation/main/MainActivity.kt
@@ -39,7 +39,7 @@ class MainActivity : AppCompatActivity() {
 
         SceytChatUIKit.chatUIFacade.channelInteractor.getTotalUnreadCount().onEach {
             binding.bottomNavigationView.getOrCreateBadge(R.id.channelsFragment).apply {
-                number = it
+                number = it.toInt()
                 isVisible = it > 0
                 maxCharacterCount = 3
                 backgroundColor = "#FA4C56".toColorInt()
